cleanup and optimization

+++ libft/src/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: okinnune <okinnune@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/29 16:06:39 by okinnune          #+#    #+#             */
-/*   Updated: 2022/03/16 15:57:18 by okinnune         ###   ########.fr       */
+/*   Updated: 2022/03/17 08:46:52 by okinnune         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,17 +16,15 @@ void	ft_bzero(void *s, size_t n)
 {
 	/*while (n > sizeof(long))
 	{
-		*(long *) s = 0L;
+		*(long *) s = (long)0;
 		s += sizeof(long);
 		n -= sizeof(long);
-		//n -= ((s += sizeof(long)) != 0) * sizeof(long);
 	}
 	while (n > sizeof(int))
 	{
 		*(int *) s = 0;
 		s += sizeof(int);
 		n -= sizeof(int);
-		//n -= ((s += sizeof(int)) != 0) * sizeof(int);
 	}*/
 	while (n-- > 0)
 		*(char *)s++ = 0;
